refactor: extract generic sidebar-header pointer-drag hook
