+++ src/renderer/src/components/sidebar/project-header-drag.ts
@@ -1,8 +1,9 @@
-import { useCallback, useEffect, useRef, useState } from 'react'
+import { useRef } from 'react'
 
 import {
   computeProjectHeaderDropPreview,
-  measureProjectHeaderDragRects
+  measureProjectHeaderDragRects,
+  type ProjectHeaderDropPreview
 } from './project-header-drop'
 import { commitProjectHeaderDragDrop } from './project-header-drag-commit'
 import {
@@ -14,303 +15,62 @@ import {
   type UseRepoHeaderDragArgs
 } from './project-header-drag-contract'
 import { createProjectHeaderDragSession } from './project-header-drag-start'
-import { getWorktreeSidebarDragAutoscroll } from './worktree-sidebar-drag-autoscroll'
+import { useSidebarHeaderPointerDrag } from './sidebar-header-pointer-drag'
 
-// Why pointer events instead of HTML5 DnD: rows are absolutely-positioned by
-// react-virtual and unmount/remount as scroll changes, so DnD enter/leave fire
-// against stale targets. With pointer events we cache the active set of repo
-// header positions and compute the drop index from the live pointer Y.
-
-export function useRepoHeaderDrag({
-  orderedRepoIds,
-  sidebarRepoHeaderIdsByBucket,
-  repoById,
-  usesProjectGroupOrdering,
-  onCommitRepoOrder,
-  onCommitProjectGroupOrder,
-  getScrollContainer
-}: UseRepoHeaderDragArgs): RepoHeaderDragController {
-  const [state, setState] = useState<RepoDragState>(INITIAL_REPO_DRAG_STATE)
-  const [sessionArmed, setSessionArmed] = useState(false)
-  const latestDropIndexRef = useRef<number | null>(null)
-  latestDropIndexRef.current = state.dropIndex
-  const orderedIdsRef = useRef(orderedRepoIds)
-  orderedIdsRef.current = orderedRepoIds
-  const sidebarRepoHeaderIdsByBucketRef = useRef(sidebarRepoHeaderIdsByBucket)
-  sidebarRepoHeaderIdsByBucketRef.current = sidebarRepoHeaderIdsByBucket
-  const repoByIdRef = useRef(repoById)
-  repoByIdRef.current = repoById
-  const usesProjectGroupOrderingRef = useRef(usesProjectGroupOrdering)
-  usesProjectGroupOrderingRef.current = usesProjectGroupOrdering
-  const onCommitRepoOrderRef = useRef(onCommitRepoOrder)
-  onCommitRepoOrderRef.current = onCommitRepoOrder
-  const onCommitProjectGroupOrderRef = useRef(onCommitProjectGroupOrder)
-  onCommitProjectGroupOrderRef.current = onCommitProjectGroupOrder
-  const getContainerRef = useRef(getScrollContainer)
-  getContainerRef.current = getScrollContainer
-  const autoscrollLastFrameTimeRef = useRef<number | null>(null)
-  const autoscrollFrameIdRef = useRef<number | null>(null)
-
-  const dragSessionRef = useRef<ProjectHeaderDragSession | null>(null)
-  const clickSwallowTimeoutRef = useRef<ReturnType<typeof setTimeout> | null>(null)
-
-  const refreshHeaderRects = useCallback(() => {
-    const container = getContainerRef.current()
-    const session = dragSessionRef.current
-    if (!container || !session) {
-      return []
-    }
-    const rects = measureProjectHeaderDragRects(container, session.bucketKey)
-    session.headerRects = rects
-    return rects
-  }, [])
-
-  const computeDrop = useCallback(
-    (pointerY: number): { dropIndex: number; dropIndicatorY: number } | null => {
-      const session = dragSessionRef.current
-      const container = getContainerRef.current()
-      if (!session || !container) {
-        return null
-      }
+export function useRepoHeaderDrag(args: UseRepoHeaderDragArgs): RepoHeaderDragController {
+  const argsRef = useRef(args)
+  argsRef.current = args
+  return useSidebarHeaderPointerDrag<
+    ProjectHeaderDragSession,
+    RepoDragState,
+    ProjectHeaderDropPreview
+  >({
+    threshold: PROJECT_HEADER_DRAG_THRESHOLD_PX,
+    initialState: INITIAL_REPO_DRAG_STATE,
+    getScrollContainer: () => argsRef.current.getScrollContainer(),
+    getSessionId: (session) => session.repoId,
+    getDraggingId: (dragState) => dragState.draggingRepoId,
+    createSession: (event, repoId) =>
+      createProjectHeaderDragSession({
+        event,
+        repoId,
+        repoById: argsRef.current.repoById,
+        sidebarRepoHeaderIdsByBucket: argsRef.current.sidebarRepoHeaderIdsByBucket,
+        getScrollContainer: argsRef.current.getScrollContainer
+      }),
+    measureRects: (container, session) => {
+      session.headerRects = measureProjectHeaderDragRects(container, session.bucketKey)
+    },
+    computeDrop: (session, container) => {
       const containerRect = container.getBoundingClientRect()
       return computeProjectHeaderDropPreview({
-        pointerY,
+        pointerY: session.latestPointerY,
         containerTop: containerRect.top,
         scrollTop: container.scrollTop,
         rects: session.headerRects,
         sidebarRepoHeaderIds: session.sidebarRepoHeaderIds
       })
     },
-    []
-  )
-
-  const cancelAutoscroll = useCallback(() => {
-    if (autoscrollFrameIdRef.current !== null) {
-      window.cancelAnimationFrame(autoscrollFrameIdRef.current)
-      autoscrollFrameIdRef.current = null
-    }
-    autoscrollLastFrameTimeRef.current = null
-  }, [])
-
-  const endDrag = useCallback(
-    (commit: boolean) => {
-      cancelAutoscroll()
-      const session = dragSessionRef.current
-      if (!session) {
-        setState(INITIAL_REPO_DRAG_STATE)
-        setSessionArmed(false)
-        return
-      }
-      try {
-        session.handleEl.releasePointerCapture(session.pointerId)
-      } catch {
-        // capture may already be released (pointercancel, element unmounted)
-      }
-      if (session.promoted) {
-        const handleEl = session.handleEl
-        const swallow = (e: MouseEvent): void => {
-          const target = e.target as Node | null
-          if (target && handleEl.contains(target)) {
-            e.stopPropagation()
-            e.preventDefault()
-          }
-          window.removeEventListener('click', swallow, true)
-        }
-        window.addEventListener('click', swallow, true)
-        clickSwallowTimeoutRef.current = setTimeout(() => {
-          window.removeEventListener('click', swallow, true)
-          clickSwallowTimeoutRef.current = null
-        }, 0)
-      }
-      const sidebarDropIndex =
-        commit && session.promoted && latestDropIndexRef.current !== null
-          ? latestDropIndexRef.current
-          : null
-      dragSessionRef.current = null
-      setState(INITIAL_REPO_DRAG_STATE)
-      setSessionArmed(false)
-      if (sidebarDropIndex === null) {
-        return
-      }
-
+    buildState: (repoId, drop) =>
+      drop
+        ? { draggingRepoId: repoId, dropIndex: drop.dropIndex, dropIndicatorY: drop.dropIndicatorY }
+        : { draggingRepoId: repoId, dropIndex: null, dropIndicatorY: null },
+    areStatesEqual: (a, b) =>
+      a.draggingRepoId === b.draggingRepoId &&
+      a.dropIndex === b.dropIndex &&
+      a.dropIndicatorY === b.dropIndicatorY,
+    commit: (session, drop) => {
       commitProjectHeaderDragDrop({
         session,
-        sidebarDropIndex,
-        orderedRepoIds: orderedIdsRef.current,
-        repoById: repoByIdRef.current,
-        usesProjectGroupOrdering: usesProjectGroupOrderingRef.current,
-        onCommitRepoOrder: onCommitRepoOrderRef.current,
-        onCommitProjectGroupOrder: onCommitProjectGroupOrderRef.current
+        sidebarDropIndex: drop.dropIndex,
+        orderedRepoIds: argsRef.current.orderedRepoIds,
+        repoById: argsRef.current.repoById,
+        usesProjectGroupOrdering: argsRef.current.usesProjectGroupOrdering,
+        onCommitRepoOrder: argsRef.current.onCommitRepoOrder,
+        onCommitProjectGroupOrder: argsRef.current.onCommitProjectGroupOrder
       })
-    },
-    [cancelAutoscroll]
-  )
-
-  const runAutoscrollFrame = useCallback(
-    (frameTime: number) => {
-      autoscrollFrameIdRef.current = null
-      const session = dragSessionRef.current
-      const container = getContainerRef.current()
-      if (!session?.promoted || !container) {
-        cancelAutoscroll()
-        return
-      }
-
-      const previousFrameTime = autoscrollLastFrameTimeRef.current ?? frameTime
-      autoscrollLastFrameTimeRef.current = frameTime
-      const autoscroll = getWorktreeSidebarDragAutoscroll({
-        point: { clientX: 0, clientY: session.latestPointerY },
-        containerRect: container.getBoundingClientRect(),
-        scrollTop: container.scrollTop,
-        scrollHeight: container.scrollHeight,
-        clientHeight: container.clientHeight,
-        elapsedMs: frameTime - previousFrameTime
-      })
-      if (autoscroll) {
-        container.scrollTop = autoscroll.scrollTop
-        refreshHeaderRects()
-      }
-
-      const drop = computeDrop(session.latestPointerY)
-      if (drop) {
-        setState((prev) =>
-          prev.dropIndex === drop.dropIndex && prev.dropIndicatorY === drop.dropIndicatorY
-            ? prev
-            : { draggingRepoId: session.repoId, ...drop }
-        )
-      }
-
-      autoscrollFrameIdRef.current = window.requestAnimationFrame(runAutoscrollFrame)
-    },
-    [cancelAutoscroll, computeDrop, refreshHeaderRects]
-  )
-
-  const ensureAutoscroll = useCallback(() => {
-    if (autoscrollFrameIdRef.current !== null) {
-      return
     }
-    autoscrollLastFrameTimeRef.current = null
-    autoscrollFrameIdRef.current = window.requestAnimationFrame(runAutoscrollFrame)
-  }, [runAutoscrollFrame])
-
-  useEffect(() => {
-    if (!sessionArmed) {
-      return
-    }
-    const onPointerMove = (e: PointerEvent): void => {
-      const session = dragSessionRef.current
-      if (!session || e.pointerId !== session.pointerId) {
-        return
-      }
-      session.latestPointerY = e.clientY
-      if (!session.promoted) {
-        const dx = e.clientX - session.startX
-        const dy = e.clientY - session.startY
-        if (
-          dx * dx + dy * dy <
-          PROJECT_HEADER_DRAG_THRESHOLD_PX * PROJECT_HEADER_DRAG_THRESHOLD_PX
-        ) {
-          return
-        }
-        session.promoted = true
-        // Why: setPointerCapture can throw if the element is detached. Check
-        // isConnected first to avoid the throw; the global pointer listeners
-        // still fire, so dragging keeps working even if capture fails.
-        if (session.handleEl.isConnected) {
-          try {
-            session.handleEl.setPointerCapture(session.pointerId)
-          } catch {
-            // Ignore capture failure; global listeners will handle the drag.
-          }
-        }
-        refreshHeaderRects()
-        setState({ draggingRepoId: session.repoId, dropIndex: null, dropIndicatorY: null })
-      }
-      refreshHeaderRects()
-      const drop = computeDrop(e.clientY)
-      if (drop) {
-        setState((prev) =>
-          prev.dropIndex === drop.dropIndex && prev.dropIndicatorY === drop.dropIndicatorY
-            ? prev
-            : { draggingRepoId: session.repoId, ...drop }
-        )
-      }
-      ensureAutoscroll()
-    }
-    const onPointerUp = (e: PointerEvent): void => {
-      const session = dragSessionRef.current
-      if (!session || e.pointerId !== session.pointerId) {
-        return
-      }
-      endDrag(true)
-    }
-    const onPointerCancel = (e: PointerEvent): void => {
-      const session = dragSessionRef.current
-      if (!session || e.pointerId !== session.pointerId) {
-        return
-      }
-      endDrag(false)
-    }
-    const onKeyDown = (e: KeyboardEvent): void => {
-      if (e.key === 'Escape') {
-        endDrag(false)
-      }
-    }
-    const onBlur = (): void => endDrag(false)
-
-    window.addEventListener('pointermove', onPointerMove)
-    window.addEventListener('pointerup', onPointerUp)
-    window.addEventListener('pointercancel', onPointerCancel)
-    window.addEventListener('keydown', onKeyDown)
-    window.addEventListener('blur', onBlur)
-    return () => {
-      window.removeEventListener('pointermove', onPointerMove)
-      window.removeEventListener('pointerup', onPointerUp)
-      window.removeEventListener('pointercancel', onPointerCancel)
-      window.removeEventListener('keydown', onKeyDown)
-      window.removeEventListener('blur', onBlur)
-      cancelAutoscroll()
-      if (clickSwallowTimeoutRef.current !== null) {
-        clearTimeout(clickSwallowTimeoutRef.current)
-        clickSwallowTimeoutRef.current = null
-      }
-    }
-  }, [cancelAutoscroll, computeDrop, endDrag, ensureAutoscroll, refreshHeaderRects, sessionArmed])
-
-  useEffect(() => {
-    if (state.draggingRepoId === null) {
-      return
-    }
-    const body = document.body
-    const prevCursor = body.style.cursor
-    const prevUserSelect = body.style.userSelect
-    body.style.cursor = 'grabbing'
-    body.style.userSelect = 'none'
-    return () => {
-      body.style.cursor = prevCursor
-      body.style.userSelect = prevUserSelect
-    }
-  }, [state.draggingRepoId])
-
-  const onHandlePointerDown = useCallback(
-    (event: React.PointerEvent<HTMLElement>, repoId: string) => {
-      const session = createProjectHeaderDragSession({
-        event,
-        repoId,
-        repoById: repoByIdRef.current,
-        sidebarRepoHeaderIdsByBucket: sidebarRepoHeaderIdsByBucketRef.current,
-        getScrollContainer: getContainerRef.current
-      })
-      if (!session) {
-        return
-      }
-      dragSessionRef.current = session
-      setSessionArmed(true)
-    },
-    []
-  )
-
-  return { state, onHandlePointerDown }
+  })
 }
 
 export {
